Move createPncAdjudicationFromAho function into root lib folder
