+++ packages/core/lib/createPncAdjudicationFromAho.test.ts
@@ -1,9 +1,9 @@
-jest.mock("./createPncAdjudication")
+jest.mock("../phase2/lib/createPncAdjudication")
 import { verdict } from "@moj-bichard7-developers/bichard7-next-data"
 
-import type { Result } from "../../types/AnnotatedHearingOutcome"
+import type { Result } from "../types/AnnotatedHearingOutcome"
 
-import createPncAdjudication from "./createPncAdjudication"
+import createPncAdjudication from "../phase2/lib/createPncAdjudication"
 import createPncAdjudicationFromAho from "./createPncAdjudicationFromAho"
 
 const mockedCreatePncAdjudication = (createPncAdjudication as jest.Mock).mockReturnValue({ dummy: "dummy" })
