+++ packages/core/lib/createPncAdjudicationFromAho.ts
@@ -1,9 +1,9 @@
-import type { Result } from "../../types/AnnotatedHearingOutcome"
-import type { PncAdjudication } from "../../types/PncQueryResult"
+import type { Result } from "../types/AnnotatedHearingOutcome"
+import type { PncAdjudication } from "../types/PncQueryResult"
 
-import { lookupPleaStatusByCjsCode, lookupVerdictByCjsCode } from "../../lib/dataLookup"
-import isRecordableResult from "../../lib/isRecordableResult"
-import createPncAdjudication from "./createPncAdjudication"
+import createPncAdjudication from "../phase2/lib/createPncAdjudication"
+import { lookupPleaStatusByCjsCode, lookupVerdictByCjsCode } from "./dataLookup"
+import isRecordableResult from "./isRecordableResult"
 
 const createPncAdjudicationFromAho = (results: Result[], hearingDate: Date): PncAdjudication | undefined => {
   const result = results.find(isRecordableResult) ?? results[0]
